projet_09_server_log_dashboard/backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.12.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
fastapi==0.109.0
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httptools==0.7.1
idna==3.11
joblib==1.5.2
kiwisolver==1.4.9
lxml==5.1.0
matplotlib==3.8.2
numpy==1.26.3
packaging==25.0
pandas==2.1.4
pillow==12.0.0
plotly==5.18.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pyparsing==3.2.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.3
requests==2.31.0
scikit-learn==1.4.0
scipy==1.16.3
seaborn==0.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==9.1.2
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uvicorn==0.27.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
